+++ hazelcast/include/hazelcast/client/mixedtype/NearCachedClientMapProxy.h
@@ -88,12 +88,12 @@ namespace hazelcast {
                                                                               long timeoutInMillis);
 
                 virtual void tryPutTransientInternal(const serialization::pimpl::Data &key,
-                                                     const serialization::pimpl::Data &value, int ttlInMillis);
+                                                     const serialization::pimpl::Data &value, int64_t ttlInMillis);
 
                 virtual std::auto_ptr<serialization::pimpl::Data>
                 putIfAbsentInternal(const serialization::pimpl::Data &keyData,
                                     const serialization::pimpl::Data &valueData,
-                                    int ttlInMillis);
+                                    int64_t ttlInMillis);
 
                 virtual bool replaceIfSameInternal(const serialization::pimpl::Data &keyData,
                                                    const serialization::pimpl::Data &valueData,
@@ -105,7 +105,7 @@ namespace hazelcast {
 
                 virtual void
                 setInternal(const serialization::pimpl::Data &keyData, const serialization::pimpl::Data &valueData,
-                            int ttlInMillis);
+                            int64_t ttlInMillis);
 
                 virtual bool evictInternal(const serialization::pimpl::Data &keyData);
